backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "arxiv>=2.4.0",
    "fastapi>=0.133.0",
    "google-api-python-client>=2.190.0",
    "google-auth-oauthlib>=1.2.4",
    "httpx>=0.28.1",
    "pdfplumber>=0.11.9",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # B008 false-positives on FastAPI Depends() in defaults

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["alembic/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
